Description
In recent years, traditional methods of pattern recognition have made great progress. However, these methods rely heavily on manual feature extraction, which can hamper the performance of the generalization model. With the growing popularity and success of deep learning methods, the use of these techniques to recognize human actions in mobile and handheld computing scenarios has attracted widespread attention. In the field of human health, Machine Learning plays a very important role in monitoring the patient's condition. In this report, I will present the theoretical basis and the construction process of a CNN and LSTM model to realize a basic human activity recognition function based on the available data. Results obtained with a high accuracy of 96.49% with the CNN model and 96.56% with the CNN-LSTM model.
As a student studying ACAD, this is a great opportunity for me to apply what I learned in school with my new knowledge of python and machine learning. This knowledge will be an excellent basis for me to enter the 5th year and for the process of working in this field at the latestIn recent years, traditional methods of pattern recognition have made great progress. However, these methods rely heavily on manual feature extraction, which can hamper the performance of the generalization model. With the growing popularity and success of deep learning methods, the use of these techniques to recognize human actions in mobile and handheld computing scenarios has attracted widespread attention. In the field of human health, Machine Learning plays a very important role in monitoring the patient's condition. In this report, I will present the theoretical basis and the construction process of a CNN and LSTM model to realize a basic human activity recognition function based on the available data. Results obtained with a high accuracy of 96.49% with the CNN model and 96.56% with the CNN-LSTM model.
As a student studying ACAD, this is a great opportunity for me to apply what I learned in school with my new knowledge of python and machine learning. This knowledge will be an excellent basis for me to enter the 5th year and for the process of working in this field at the latest